Issue #377 - use the VMUSB drivers in place rather than as copies
Use V1729 support from vmusb/daqconfig directly rather than from
an  unmaintainable local copy.
